requirements.txt
asgiref==3.11.1
channels==4.3.2
daphne==4.2.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.7.0
PyJWT==2.11.0
python-dotenv==1.2.1
scikit-learn==1.7.2
sentence-transformers==3.4.1; python_version < "3.14"
pypdf==6.1.1
python-docx==1.2.0
sqlparse==0.5.5
celery==5.4.0
redis==5.0.8
psycopg[binary]==3.2.12
